+++ examples/restful/margin/getAllMarginAssets.ts
@@ -4,9 +4,14 @@ import { RestMarginTypes, Spot } from '../../../src/index';
 dotenv.config();
 
 const apiKey = process.env.BINANCE_API_KEY || '';
+const apiSecret = process.env.BINANCE_API_SECRET || '';
 const baseURL = process.env.BINANCE_BASE_URL || '';
-const client = new Spot(apiKey, '', { baseURL: baseURL });
+const client = new Spot(apiKey, apiSecret, { baseURL: baseURL });
 
-client.getAllMarginAssets().then((res: RestMarginTypes.getAllMarginAssetsResponse[]) => {
+const options = {
+    asset: 'BNB',
+};
+
+client.getAllMarginAssets(options).then((res: RestMarginTypes.getAllMarginAssetsResponse[]) => {
     console.log(res);
 }).catch(err => { console.log(err); });